requirements.txt
-f ../../../prebuilt

bosdyn-client >= 3.1
six
pyqt5
Pillow==6.1.0
windows-curses==2.1.0; sys_platform == 'win32'